Explicitly flush when writing access log entries
daphne 1.0.0 arranged to flush access logs immediately (#53, #54), but
only when logging to a file.  It's often convenient to log to stdout
instead and rely on the systemd journal, but at the moment that results
in access logs being unhelpfully block-buffered.  Explicitly flush
entries so that this works regardless of the logging destination.